feat: 完成 java_mybatis 12 项改进并全量回归通过
(1) 转账修复——金额正数校验（负数可反向造钱）、AccountMapper 改原子扣减
    debit/credit（UPDATE ... balance=balance-? WHERE id=? AND balance>=?）
    + 影响行数判定防并发超扣
(2) Account.balance double→BigDecimal（含实体/接口/Demo/测试，断言用 compareTo）
(3) 测试隔离——app.db.url 系统属性注入 mybatis-config 的  占位符，
    surefire 固定测试用 target/test.db
(4) AccountServiceTest 每用例 @Before 重建库，彻底去除 @FixMethodOrder 顺序敏感
(5) 新增 CacheTest 验证二级缓存命中(assertSame)/insert 失效
(6) 抽 SqlSessionTemplate 统一 session 生命周期（execute 自动提交 / executeInTransaction
    程序化事务）
(7) 新增 BizException 统一业务异常
(8) logImpl=SLF4J + org.apache.ibatis DEBUG 打开 SQL 日志
(9) CI maven.yml 去掉 stale MySQL 改 SQLite+JDK24
(10) README 重写 + AGENTS.md 同步
(11) docs/spring-integration-guide.md 交付 Spring 演进指南
(12) 密码哈希以 README 注明教学演示处理

26 测试 0 失败、Checkstyle+SpotBugs 0 缺陷

+++ src/main/java/com/example/service/AccountService.java
@@ -1,5 +1,13 @@
 package com.example.service;
 
+import java.math.BigDecimal;
+
 public interface AccountService {
-    void transfer(Integer fromId, Integer toId, double amount);
+
+    /**
+     * 转账：fromId 账户扣减 amount，toId 账户增加 amount。
+     * 金额必须为正数，否则抛 {@link com.example.exception.BizException}；
+     * 任何一步失败（余额不足、账户不存在）整个事务回滚。
+     */
+    void transfer(Integer fromId, Integer toId, BigDecimal amount);
 }
